Refactor package structure for BoundArmorAnimator and TaggedArmorAnimator to improve organization
